requirements.txt
python-dotenv==1.0.0
langchain-openai==0.0.2
openai>=1.10.0
playwright>=1.35.0
langchain>=0.0.311
pydantic>=2.0.0
asyncio>=3.4.3
typing-extensions>=4.5.0
gradio>=4.0.0 